pyproject.toml
[project]
name = "self-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.95.0",
    "dspy>=2.6.8",
    "google-genai>=1.3.0",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.3.5",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.11",
    "langgraph>=0.3.20",
    "ollama>=0.4.7",
    "python-dotenv>=1.0.1",
]
